[scava2es] Handle configuration smells data
This code allows to insert configuration smells data into the ElasticSearch
database. The current enriched items contain the smell name, the line and
file name where the smell was detected, the reason, the type of configuration
(puppet or docker), the smell type (implementation or design) and the sub
type (e.g., antipattern), some optional attributes such as the commit and
date when the smell was found (`date`), plus information about
the project (`project` and `meta`), when the project was updated, the id
of the smell and the uuid of the enriched item. It is worth noting that
the field `datetime` is derived using the `date` attribute (if not None)
or the project `updated` attribute. An example of enriched item is shown below:

          "smell_name" : "Inconsistent naming convention",
          "line" : " 24",
          "reason" : " python::pyvenv not in autoload module layout ",
          "file_name" : "/root/scava/puppetpython/.../manifests/pyvenv.pp ",
          "commit" : "c3bb6b37e7711f816faba3ea8fb98b72286830e1",
          "date" : "2019-05-21T18:40:17.000Z",
          "conf_type" : "puppet",
          "smell_type" : "implementation",
          "smell_sub_type" : "antipattern",
          "id" : "puppet_implementation_Inconsistent...",
          "updated" : "20190402",
          "project" : "puppetpython",
          "datetime" : "2019-05-21T18:40:17+00:00",
          "uuid" : "11eb644184b9a752ccda2e00af5a76d9bfbcc32b",
          "meta" : {
            "top_projects" : [
              "main"
            ]
          }