[dashboard] Add initial version smells devops dashboard
This code proposes an initial version of the smells devops
dashboards. It includes two visualizations to focus on a
specific project and/or top project. Two bar charts show
the evolution of the smells extracted from puppet and
docker. They rely on the following historic metrics:
- puppet.designsmells.smells
- puppet.implementation.smells
- docker.smells

Two tables allow to focus on the smells details. The
former shows information about design and configuration smells,
while the latter gives insights about antipatterns smells.