[scava2es] Handle commits topics data
This code allows to process commits topics data and insert to ElasticSearch.
The current enriched items contain the topic name, the date (i.e., `date`
and `datetime`) when the metric was calculated, the number of comments for
that topic, the scava metric from where the data was fetched, project and
meta project information.

          "topic" : "Wider Community",
          "date" : "20190307",
          "comments" : 2,
          "scava_metric" : "bugs.topics.comments",
          "updated" : "20180430",
          "project" : "microprofile",
          "datetime" : "2019-03-07T00:00:00+00:00",
          "uuid" : "6e3b802219e41792d7c3a0744b3305c42f1e249c",
          "meta" : {
            "top_projects" : [
              "main"
            ]
          }