[scava2es] Remove scava data from ES items
This code removes the scava data from the ES items, thus
reducing the size of the indexes. Furthermore the metric
`documentation.readability.entries` is filtered out due to
a wrong format (it's not a linechart) which breaks the
scava2es importer process. Finally the metric
`Method Hiding Factor (Java)` is added again together with
the modified version `Method Hiding Factor (Java) GT 10`.