[web-dashboards] Shorten legends for visualizations
This code shortens the legends of some visualizations
in the following dashboards:
- scava-overview
- scava-sentiment-emotion
- scava-users

Furthermore, scava-qm includes a new visualization that
shows the average scores of goals based on the average
scores of attributes. Thus, it completes in this way
the original visualization that shows the average
scores of goals based on the average scores of qm metrics.